move dummy_field_adapter to adapter folder
